requirements.txt
Flask==3.0.3
Flask-Mail==0.9.1
mysql-connector-python==9.0.0
Flask-Login==0.6.3
itsdangerous==2.2.0
Werkzeug==3.0.3
reportlab==4.2.2
python-dotenv==1.0.1
gunicorn==21.2.0
qrcode==7.4.2
mailersend==0.5.4 
Pillow==10.1.0
